Rename new .vue components.
