+++ client/galaxy/scripts/components/Message.vue
